Change "service_data" to "data"
AFAIK, service_data is not accepted syntax anywhere.. It took me quite a while to abandon the suggested syntax and just change it to "data" as it should be anywhere where you'd actually use this.. I think it would save people some time if usable syntax was shown. But maybe I'm missing something.