Allow pop-up cards to be opened while hidden.
Setting the pop-up card element's hidden attribute completely removes
it from the layout and prevents unsightly gaps in section views.
Unfortunately, that also disconnects the element from the document
so we can't use the connected callback to determine when to register
the event handlers.

To work around this issue, globally register the more-info event
handler on the window and search the Lovelace configuration for
the card definition so it doesn't matter whether the card is visible
and it can even be defined in any view of the same dashboard.